implement user and auth routes, authentication system and JWT

+++ backend/src/app/domain/services/user_service.py
@@ -1,5 +1,6 @@
 from uuid import UUID
 
+from src.app.core.exceptions.base_exc import NotFoundError
 from src.app.core.exceptions.user_exc import UserAlreadyExists
 from src.app.core.security.auth_manager import AuthManager
 from src.app.domain.models.db.user import User
@@ -30,7 +31,11 @@ class UserService:
         user = await self.repository.get(id=id)
 
         if not user:
-            return None
+            raise NotFoundError(
+                entity_type_str='User',
+                field_name='id',
+                field_value=id
+            )
 
         return user.to_dto()
 
@@ -42,6 +47,13 @@ class UserService:
         :return: UserDTO representing a user.
         """
         user = await self.repository.get_by_username(username=username)
+
+        if not user:
+            raise NotFoundError(
+                entity_type_str='User',
+                field_name='username',
+                field_value=username
+            )
 
         return user.to_dto()
 
@@ -63,7 +75,7 @@ class UserService:
 
         :return: DTO representation of created User
         """
-        user = await self.get_by_username(username=schema.username)
+        user = await self.repository.get_by_username(username=schema.username)
 
         if user:
             raise UserAlreadyExists(username=user.username)
@@ -88,8 +100,17 @@ class UserService:
 
         :param id: user id
 
-        :return: ``True`` if user was deleted
+        :return: True if user was deleted
         """
+        user = self.repository.get(id=id)
+
+        if not user:
+            raise NotFoundError(
+                entity_type_str='User',
+                field_name='id',
+                field_value=id
+            )
+
         deleted = await self.repository.delete(id=id)
         await self.repository.session.commit()
